requirements.txt
aiogram>=3.4.0
python-dotenv>=1.0.0
transformers>=4.40.0
torch>=2.1.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
xgboost>=2.0.0
catboost>=1.2.0
joblib>=1.3.0
